AI Career Navigator/requirements.txt
google-genai==1.36.0
streamlit==1.49.0